+++ test/libsolidity/smtCheckerTests/functions/recursive_multi_return.sol
@@ -9,5 +9,3 @@ contract C {
 // ====
 // SMTEngine: all
 // ----
-// Warning 6321: (48-52): Unnamed return variable can remain unassigned. Add an explicit return with value to all non-reverting code paths or name the variable.
-// Warning 6321: (54-58): Unnamed return variable can remain unassigned. Add an explicit return with value to all non-reverting code paths or name the variable.